requirements.txt
flask>=3.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
